Refactor folder and add unit tests for vector and rotation
